Rename gold file so it doesn't conflict with pytest commandline arg
