+++ sam/sim/test/gold.py
